requirements.txt
alembic==1.13.1
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
Flask==3.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
Mako==1.3.0
MarkupSafe==2.1.4
oauthlib==3.2.2
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simple-websocket==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
